CMS-12160 Make all links relative
The cms can be behind a reverse proxy or load balancer. A reverse proxy
maps the external url to the url of the cms war running in tomcat.
For instance: https://example.com/foo/ => http://10.10.100.121:8080/cms
When a request is forwarded by the proxy to tomcat then it adds some
http headers:
X-Forwarder-Scheme (https)
X-Forwarded-Host (example.com).
Unfortunately the part "foo" is not sent as a header, so inside the CMS
application we are unable to reconstruct the full URL.

Since the navapp must call a CMS REST service to get the nav items we
need a new kind of ResourceType to indicate that the resource.url is
an incomplete url. The navapp will have to prefix it with the navapp
"home" url. In the example above, this "home" url would be
https://example.com/foo/