CMS-12155 for action requests, fallback to Referer if Origin header misses
If the Origin header is missing, the OWASP recommendation is to check
the Referer header as fallback. If the Referer header is also missing,
the recommendation is to abort the request, which we now do. Note that
we do not support configuration to allow missing Origin and missing
Referer. The Referer must be present as a minimal fallback.

Also see from OWASP:

Identifying Source Origin (via Origin/Referer header)
Checking the Origin Header

If the Origin header is present, verify that its value matches the
target origin. Unlike the Referer, the Origin header will be present
in HTTP requests that originate from an HTTPS URL.

Checking the Referer Header

If the Origin header is not present, verify the hostname in the
Referer header matches the target origin. This method of CSRF mitigation
is also commonly used with unauthenticated requests, such as requests
made prior to establishing a session state, which is required to keep
track of a synchronization token.

In both cases, make sure the target origin check is strong. For example,
if your site is site.com make sure site.com.attacker.com does not pass
your origin check (i.e., match through the trailing / after the origin
to make sure you are matching against the entire origin).

If neither of these headers are present, you can either accept or
block the request. We recommend blocking. Alternatively, you might want
to log all such instances, monitor their use cases/behavior, and then
start blocking requests only after you get enough confidence