REPO-2239 Check the role hippo:editor instead of jcr:write on folder
The role hippo:editor does not imply jcr:write any more (as it shouldn't),
and thus an editor does not have by default jcr:write on folders.

Whether a cms user can copy to a certain destination should be concluded
from whether (s)he has the rolee hippo:editor on the destination. The
actual copy is on the end done by a workflow session which is allowed
to do that and has nothing to do with whether the cms user has write
access to a folder