REPO-2239 Check hippo:author role instead of JCR_WRITE
Hippo authors do not have write access on folders any more, but should
be able to read the gallerytypes. The original check on write access
was a bit strange to begin with (authors had write access to all folders).

Instead, we now check on the role 'hippo:author' which is a better role
to check since the logic for 'gallerytypes' is more a functional role
than a jcr permission