CMS-12093 Add redirect filter
This filter will redirect to the home page of the cms and add a
query parameter initialPath if the http request
- has method GET
- does not have an iframe query parameter
- does not start with a whitelisted prefix
otherwise it will pass the request and response down the filter chain.

We need this to support deep linking to paths that are unknown inside
the CMS (e.g. paths to brSM apps).