REPO-2242 Support for federated domain security
This is a manually squashed, and further optimized, feature created by Ard Schrijvers!

Federated domains are located below nodes of type hipposys:federateddomainfolder, which extends hipposys:domainfolder,
and are always sandboxed relative to the parent of the hipposys:federateddomainfolder.
A hipposys:federateddomainfolder therefore may not be located directly under the jcr:root, and neither under /hippo:configuration.

Standard domains are now only allowed under /hippo:configuration/hippo:domains which must be of primary type hipposys:domainfolder.
This also obsoletes the usage/need for the /hippo:configuration/hippo:security/hipposys:domainspath property!

Domains under a not-standard/not-supported domainfolder location will be *ignored* (with a warning logged) by the SecurityManager!

Because domains now can be located under multiple folders, and the SecurityManager loads and processes them in a set, the
equals method of domains has been changed to compare their node path.