CMS-12202 Add and use user roles xm-cms-user and xm-console-user for login check
This replaces (and removes) the current cms/console application login check requiring to
configure a specific repository privilege which already authenticated users were required to have
at a specific (but actually arbritrary) configured repository path.

This typically was a cumbersome setup if/when needing to customize it.

Instead, by introducing the user roles xm-console-user and xm-cms-user, this now becomes really
trivial and no longer needs an arbritary privilege configuration through security domains.

The checkPermission method in the IPluginConfigService, which was *only* used for this one purpose
no longer is needed and has been dropped from the interface / implementations.