NAVAPP-161 Add loginmessage key
After logout the browser is redirected to the CMS. The CMS will show
the login screen because the user has just been logged out. The server
will lookup the loginmessage key to display a message on the login
screen.

Note:
I used the DOCUMENT injection token instead of the location service
because the location service can only deal with internal routing.
What we need instead is a "hard" reload outside of the angular router.
The benefit of using document.location over window.location is that
the DOCUMENT can be injected as a mock for testing.