REPO-2244 get rid of the usage/need for an anonymous jcr session
Practically we only used an anonymous jcr session for the CMS login screen to read its needed configuration.
In a following commit this will be replaced with a dedicated (jvm) system session ("frontend-system-user").

By dropping the need for an anonymous session, a lot of additional code cleanup became possible as well,
as now there only remains two types of jcr sessions: either a system session or a (regular) user session.

Therefore, we no longer need using additional Subject principals to cater for the different combinations
(an anonymous session still wold be part of the "everybody" group (* matcher), and thus needed separate
GroupPrincipal(s), as well as the FacetAuthPrincipal to hold the derived FacetAuthDomains.

Now, all those variations collapse and need only a UserPrincipal, which already (also) did hold the
groupIds (getUser().getMemberships()), and now has been extended to also hold the FacetAuthDomains.
This removes the need for both the GroupPrincipal and FacetAuthPrincipal, which therefore have been dropped
(as well as the already not used RolePrincipal).

Verification that anonymous login or impersonation as anonymous user no longer is possible is provided
through the unit tests in AnonymousUserTest