REPO-2244 use a dedicated (jvm) system session as 'fallback' session before a user has logged in.
Currently this system session ("frontend-system-user") retains the same privileges as the previously
used anonymous jcr session, granted through its membership of the "everybody" group (* matcher),
and the "defaultread" domain granting readonly access to the "everbody" group members.

In the future, we can (now) further restrict the needed defaultread access currently needed for the
frontend-system-user by *moving* specific defaultread domain rules to a dedicated domain to be
granted to that frontend-system-user.