CMS-12197 Register sub app iframes
Register sub app iframes, so that calls to those iframe can only be made
once the connection promise has been fulfilled.

Add jasmin to engine module (thanks to Arthur and Joeri)

Use the iframes as keys to the connection map, so they can be looked up
by the classname of the perspective.