REPO-2243 Force end projects to fail bootstrapping in case not upgraded
Pre 14.0.0 the security domain for 'live-documents' (and 'preview') was
called 'live-documents'. If we keep it this name, then typically an
end project bootstraps the group 'liveusers' to it, and to the group
'liveusers' they bootstrap the liveuser, and potentially another liveuser,
eg 'intranetliveuser'

This bootstrapping keeps working, and gives the 'intranetliveuser' user
read acces to live documents. *However*, 'intranetliveuser' should
instead get the userrole 'xm-hst-live-documents-reader'

To force end projects to fail if they did not take the right upgrade
actions, we rename the domain name to force failure