CMS-12209 Give focused fields a blue border
The actual border is almost always drawn around a parent element, but
we can use the :focus-within selector to recognize a focused field
(except on Edge, according to the Internet, but that may change when
Edge runs on Chromium too).

We want  2px blue border, but the normal border is only 1px. Setting the
border itself to 2px pushes content down, which is not so nice. The
solution is to use box-shadow for the additional 1px blue border, which
does not affect the layout :).