REPO-2256 Assert which property types are fully supported in facet rules
Boolean, Long and Double are also supported in the authorization query
and thus can be fully utilized in FacetRule.

Date and BigDecimal are not support in the authorization query and are
as a result not fully supported in a FacetRule. They *could* theoretically
be used for providing write access to nodes for which read access is
already covered