HSTTWO-4652 Support privilege based channel admin checking
The check for the property

/hippo:configuration/hippo:frontend/cms/hippo-channel-manager/
    channel-manager-perspective/templatecomposer/@manage.changes.privileges
has become completely redundant.

Instead, there has been introduced a new privilege and a new role for it:

/hippo:configuration/hippo:roles:
  /channel-admin:
    jcr:primaryType: hipposys:role
    hipposys:privileges: ['hippo:channelAdmin']
    hipposys:roles: [readwrite]

A cms user is a channelAdmin (can manage changes of others and
potentially delete a channel) on a specific channel if it has the
privilege hippo:channelAdmin on the channel configuration, thus for
example on

/hst:myproject/hst:configurations/myproject

If there is a preview, the cms user is only channelAdmin if it then
also has privilege hippo:channelAdmin on the preview configuration.

Note that you can now have a user that is channelAdmin on channel A but
not on channel B. This has been changed compared to before, since back
then the check was done against /hst:hst only (and by default if you
had hippo:admin there)

Next to that, in the hst:domains, we now also bootstrap the
following authrole

/channel-admin:
  jcr:primaryType: hipposys:authrole
  hipposys:groups: []
  hipposys:role: channel-admin
  hipposys:users: []

as can be seen, nobody is by default a channelAdmin any more, not even
the admin user. Making someone a channelAdmin is as simple as adding
the user to hipposys:users and likewise for an entire group to hipposys:groups

Since we now really check against a JCR privilege and the check in the
jaxrs resources was done against RolesAllowed,
I've introduced a new annotation called PrivilegesAllowed. This
annotation can be used to protect JAXRS resource methods (endpoints)
by requiring a certain privilege.

This mechanism is much more flexible, and can be used for other
privileges than 'hippo:channelAdmin' as well