NAVAPP-161 Return parentApi in then
Should not be needed if I understand the Promise API specification
correctly. I pass in something that is not a function, so I expected
the promise to return the parentApi.
Apparently it does not! In the javascript console this error is logged:
TypeError: Cannot read property 'updateNavLocation' of undefined

I can fix this issue by returning the parentApi (so making the handler
a function).