HSTTWO-4653 Improve Channel Manager privileges support
In the channel manager we have currently the following different roles

1) channel admins who are allowed to manage others' changes and if
configured are allowed to delete a channel
2) channel webmasters who are allowed to modify the hst configuration
and publish their own changes
3) channel 'webviewers' who are allowed to view the channel preview,
and for example possibly be editor or author on documents

We should better cater for these different roles in the hst security
configuration via privileges, and also make sure that methods are only
invoked when the right privileges are present. We should set up
something that in the future can also cater for extra features, like
more enhanced workflow. For example introduce later on 'channel
publishers' which is the required privilege to publish a channel,
just to name something.

To achieve the setup described above, we introduce the following hippo:roles :

    /hippo:configuration/hippo:roles:
      /channel-admin:
        jcr:primaryType: hipposys:role
        hipposys:privileges: ['hippo:channel-admin']
        hipposys:roles: [channel-webmaster]
      /channel-webmaster:
        jcr:primaryType: hipposys:role
        hipposys:privileges: ['hippo:channel-webmaster']
        hipposys:roles: [readwrite, channel-webviewer]
      /channel-webviewer:
        jcr:primaryType: hipposys:role
        hipposys:privileges: ['hippo:channel-webviewer']

Note that channel-admin > channel-webmaster > channel-webviewer.
This a channel-adkin is also a channel-webmaster and channel-webviewer.

Next to the roles above, we do something similar for userroles
(again with similar inheritance):

/xm-channel-admin:
        jcr:primaryType: hipposys:userrole
        hipposys:userroles: [xm-channel-webmaster]
        hipposys:system: true
      /xm-channel-webmaster:
        jcr:primaryType: hipposys:userrole
        hipposys:userroles: [xm-channel-webviewer]
        hipposys:system: true
      /xm-channel-webviewer:
        jcr:primaryType: hipposys:userrole
        hipposys:system: true

Finally, we make sure that the group 'editor' and 'author' by default
have the userrole 'xm-channel-webviewer', and that the group 'webmaster'
has the userrole 'xm-channel-webmaster'.

For hte hst:domains, we specify three authroles, namely:

          /channel-admin:
            jcr:primaryType: hipposys:authrole
            hipposys:role: channel-admin
            hipposys:userrole: xm-channel-admin
            hipposys:groups: []
            hipposys:users: []
          /channel-webmaster:
            hipposys:role: channel-webmaster
            jcr:primaryType: hipposys:authrole
            hipposys:userrole: xm-channel-webmaster
            hipposys:groups: []
            hipposys:users: []
          /channel-webviewer:
            hipposys:role: channel-web-author
            jcr:primaryType: hipposys:authrole
            hipposys:userrole: xm-channel-webviewer
            hipposys:groups: []
            hipposys:users: []
Now we have all the bits and bytes to make sure that

1) channel admins have the privileges hippo:channel-admin,
hippo:channel-webmaster and hippo:channel-webviewer
2) channel webmasters have the privileges hippo:channel-webmaster
and hippo:channel-webviewer
3) channel webviewers (editors/authors) have the privilege
hippo:channel-webviewer

With the new support for @PrivilegesAllowed from HSTTWO_4652,
we can now protect every rest endpoint to only be allowed to be invoked
by a user with enough privileges.

Note that in the past invoking a method you were not allowed to invoke
did not have real effect because an author did for example not have
write access on the hst config nodes. However, now it is much cleaner
and what we now can achieve in the future is that all the actual JCR
node changes are done by a workflowuser and we remove the readwrite
from the channel-webmaster role : namely, the actual changes can be done
by workflow, the privilege check is done for checking whether the
invocation is allowed. This can now become very similar to how document
workflow is done where the changes are persisted through a workflow
session