HSTTWO-4653 revert last changes in PrivilegesAllowedInvokerPreprocessor
With commit #c7284cb I made a mistake I now realize. It is
a subtle but important one.

PrivilegesAllowed means that it is already allowed
if just 1 privilege matches. However my change in commit
#c7284cb resulted in that *every* privilege should
match. This would thus be more like PrivilegesRequired

However, since it is PrivilegesAllowed, I cannot use
the simple #hasPrivileges but really need to find the
intersection in the more verbose way