NAVAPP-253 Implement user activity handling
When the active app reports user activity the parent will invoke
onUserActivity on all inactive apps that implement this API method.
To prevent flooding apps with these calls we throttle by one minute.

NOTE:
because users can't do anything in inactive apps it is assumed that only
the active app invokes onUserActivity on the parent.