HSTTWO-4652 @PermitAll annotation and make access annotation mandatory
Some rest endpoints are invoked when there is not yet a channel
selected, meaning that the channel against which the privilege needs to
be checked is unknown. Hence, we need the @PermitAll for these kind
of methods.

Also note that from now on, a rest endpoint MUST have a security
annotation, thus either PermitAll OR PrivilegesAllwed. If both missing,
the method is not allowed. If both present, it is neither allowed and
an error is logged