HSTTWO-4652 Only return allowed channels on #getChannels
Returning channels for which the cms user is not at least channel-webviewer
doesn't make sense. Hence, by default only return channels for which the
cms user is channel-webviewer. For something like cross channel page
copy, the user should have the channel-webmaster privilege on the
returned channels : Otherwise the user is not allowed to make modifications
on the target channel