REPO-2209 Make hipposys:filter optional
Most of the time, the hipposys:filter is only confusing when present
in a facet rule. Only for very specific facet rules, the
hipposys:filter must be set with value true, but in 90% of the facet
rules, it is present with value false. But in case of something like a
NodeName or NodeType match, hipposys:filter (optionality) is meaningless

Hence make it non-mandatory and when missing, it just means

hipposys:filter = false