CHANNELMGR-2414 Update nav location whenever active channel changes
I tried to implement logic to call updateNavLocation in the parent
iframe communication service but that lead to circular references.

Since the channel service manages the active channel it makes sense to
connect to the CMS (the parent) in that service. Still had to make
ChannelService.updateNavLocation() public. This is needed because the
channel controller needs to update the nav location when a channel is
opened from the content perspective. If the channel for the document is
already open, then the channel is not loaded, so a call to update the
nav location is required.