HSTTWO-4658 Extend PrivilegesAllowed to optionally have an absPath
When an absPath is provided, the privileges on that absolute path will
be checked instead of 'some context aware' jcr path. If there is no
absPath, then a subclass can check a context aware location. By default
the PrivilegesAllowedInvokerPreprocessor only uses the absPath and if
no absPath is present, returns that the method is not allowed. A subclass
can in case of a missing absPath check against some context aware path,
for example the path of the Channel configuration that is currently
stored in some context (say http session)