NAVAPP-188 Invalidate cache for filelist.json
Invalidate the cache for filelist.json by changing an antiCache query
parameter if a newer version of the navapp is deployed so users don't
have to do a 'hard refresh' anymore.  The value is the same value as the
the antiCache parameter of loader.js.  This is only relevant if the
loader is served by brX. A CDN manages its own cache invalidation.  The
cache is invalidated if any of the jars of the cms war have changed.
The jar of the "engine" module of the hippo-cms project contains the
navapp distribution. If any file of the navapp distribution changes, the
antiCache parameter of loader.js changes and thus the antiCache
parameter of filelist.json.