HSTTWO-4662 hst-addon-repository should pull in hst-addon-cnd
Currently downstream projects which want to bootstrap

<dependency>
  <groupId>org.onehippo.cms7.hst.toolkit-resources.addon</groupId>
  <artifactId>hst-addon-repository</artifactId>
</dependency>

always need to include hst-addon-cnd as well if not relying on the
'platform-dependencies' pom module. There is no reason why the
'hst-addon-repository' would not pull in the hst cnd since it
is always needed.

Hence now transitively pull in the dependency below in 'hst-addon-repository'

<dependency>
  <groupId>org.onehippo.cms7.hst.toolkit-resources.addon.toolkit-cnd</groupId>
   <artifactId>hst-addon-cnd</artifactId>
</dependency>