REPO-1971 require the path (or parent path) to exist for permissions/privileges checks
This refactoring actually enforces and simplifies the handling of two requirements:
- Standard JCR modifying operations (JCR_ACTIONS or JCR_WRITE_PRIVILEGES) can only be allowed when having at least read access
- Given the previous requirement it now becomes possible to enforce a path (or parent path) to exists for a permission/privilege check

However for the latter requirement there is one exception: when checking the Session.REMOVE action (for remove node and/or remove property).
Because of a quirk or limitation in Jackrabbit, there are specific scenarios when the path to check that permission against may not exist (anymore).
So a Session.REMOVE check for a path which doesn't exist (still) returns true/allowed.
In all other cases a PathNotFoundException will be thrown if the path to check does not exist.

For more details, see https://issues.onehippo.com/browse/REPO-1971