NAVAPP-266 Fix initial navigation bug caused by lacking of the active app
The lack of the active app is the case during initial navigation because it's set after successful navigation. If the app calls parentApi.udpateNavLocation() during childApi.navigate() call it leads to interaction with the navapp which not fully initialized. To solve this issue the active app is set right before the childApi.navigate() method call which makes it logical because all necessary information is already there and error will be shown if something does wrong.