CHANNELMGR-2419 Reload channel on navigate
Reload the channel when a navigate call is received by the channel
manager so that cluster wide changes are applied.

Introduce navigate method, so that the methods object that is send over
http5 postmessage is as small as possible in order to limit the possible
of errors during marchalling and unmarchalling json.