CMS-12065 Adding userroles assignment management to the Users plugin/panel for the CMS admin perspective
Userroles now can be added and removed for all users (internal and external).

This enhancement now also restricts to readonly access for users with (only) the xm-security-application-manager userrole,
while providing crud operations for users with the xm-security-user-manager userrole.

Also, the practically not usable 'feature' to set/change the security provider (introduced with CMS-8085) has been
removed.