CMS-12297 Do not navigate via channel editor iframe
The channel manager perspective is kind of a special case. It is a
mixture of Ext and an angular  app in an iframe. There is no reliable
way to determine when the angular app in the iframe is initialized.
Therefore we shouldn't invoke navigation methods on it. Fortunately we
do not have to do so because we do not support deep linking via the
navapp to a specific channel. Navigation to a channel is handled
internally inside the CMS and does not involve the navapp.