REPO-2267 Introduce the role xm-content-holder
Anybody who is holder of a content item is allowed to write to it,
regardless where in the repository the content item is located. This is
a domain that is universal and never needs to be changed. Since it holds
everywhere and for everybody, the group 'everyboddy' now has the
userrole 'xm-content-holder'.

Since gallery and asset items do not have a holder since there is no
workflow on those items, a separate domain has been created for
gallery and assets.