CMS-12299 Call navapp from hidden AjaxIndicator
Include AjaxIndicator again, but his time hidden, so it can call the
navapp to show or hide the "BusyIndicator".

Only call hide if the wait class is actually there to prevent unbalanced
show/hide calls. The navigation application expects a hide call for each
show call.

Bump navapp and navapp-communication to be able to use the
hideBusyIndicator and showBusyIndicator calls.