REPO-2269 hippo:facetselect is not a virtual node provider anymore
When we have the following node structure:

+ mydoc (hippo:handle)
   + mydoc (hippo:document)
      + html (hippostd:html)
            + link (hippo:facetselect)

then if the user has

1) read access to mydoc/mydoc/html/link
2) write access to mydoc/mydoc

then the user should also have write access on mydoc/mydoc/html/link,
see REPO_2236

However, only nodes of type hippo:facetselect this does not work since
for facetselect nodes it is skipped to test whether there is an
ancestor that is a document. The reason
for this does not exist anymore: It was like this when facetselect was
still a virtual node provider (just like mirror), which it isn't any
more since many years

Note that there is a comment that the jcr session needs to be refreshed
after a change on a node of type mirror or facetselect. This is a
separate bug that needs to be addressed separately