REPO-2271 Introducing a new DomainsManager provided through the RepositorySecurityManager
The DomainsManager provides a replacement for needing direct JCR read/write access for a
xm-security-application-manager userrole granted administrator (e.g. for usage in CMS Admin).
Those operations are now handled through the internal system session of the DomainsManager,
and we no longer need a dedicated (and complicated to setup/maintain) security domain for it.

Although the operations provided by the DomainsManager are currently limited to basic crud for hipposys:authrole nodes,
it covers the use-cases for the current CMS Admin (Permissions) plugin.
The corresponding changes to adapt and use the DomainsManager for those use-cases will be done through CMS-12065