HSTTWO-4671 [Backport 13.4] Move finding of parent bean to HippoContentBean
Parent of a document type will be assigned on HippoContentBean.
If a supertype exists in the same namespace, then that will be used to
get the parent bean. If there isn't any matching supertype in the same
namespace, then HippoDocument and HippoCompound supertypes
will be checked on the supertypes. If one of them exists, then that
will be used as parent bean. If none of them exists, then no parent
bean will be assigned to HippoContentBean.

(cherry picked from commit a4b62a7f6d29e46d34d4fd861235622e8bceed7d)