CMS-12313 Support for plugin instantiation failure
If a plugin during instantiation 'discovers/finds out' that it is not
correct, for example because the user jcr session does not have the
right privilege on some jcr node, the plugin should be able to be
short-circuited.

What still needs to be done, is that the plugin that fails should be
removed from the parent plugin if that is possible. For example if
the user does not have read access to 'gallery' items, then the item
'Gallery' should be removed from the drop down list in the content
perspective. This still has to be done.