HSTTWO-4670 Provide new RepositoryAuthenticationProvider based on userroles
The replacement for the now deprecated JcrAuthenticationProvider/HippoAuthenticationProvider implementations of AuthenticationProvider
make (only) use of filters/maps SessionUser.getUserRoles() to produce the set of (hst) Role principals of an authenticated user.

The javadoc already explains in details all the configuration options for the RepositoryAuthenticationProvider, but most importantly
it no longer needs jcr read/query access to security users/groups/domains (as the two above do).
As a repository authenticated user (HippoSession) already has a fully detachable SessionUser object, with all the resolved userroles
(and groups, although not used here), all data needed for this much more efficient implementation is already immediately available.

All the existing capabilities of the JcrAuthenticationProvider/HippoAuthenticationProvider are still present, plus some additional
new ones, like (optionally) requiring a specied userrole for the user to allow the authentication (and/or requesting its userroles).

The above mentioned deprecated implementations are still fully functional and supported, but will be removed in v15+.
Also, the HippoAuthenticationProvider is still kept the 'default', but this will be replaced in a follow-up commit to become the
new RepositoryAuthenticationProvider.