HSTTWO-4672 Channel can be missing or previously stored channel is present
When switching to another channel, the first couple of requests still
have the previous channel on the http session. Hence the
InvokerPreprocessor instances can have a reference to the wrong channel
and for example do a check (is the channel locked) against the wrong
channel.

Typically, requests that are used to switch to a new channel (for example
the current channel is 'myproject' and a ajax request is done for
/channels/subproject to fetch information about the subproject, then
that method should not do a check against 'myproject' to test for example
whether the hst configuration is locked.

Hence introduce a new annotation, @ChannelAgnostic, to be able to add to
methods which are channel agnostic and thus for example never need the
HstconfigLockedCheckInvokerPreprocessor to check the channel being locked

The @ChannelAgnostic annotation is now added to all methods involved in
switching to a different channel or fetching a certain channel info