REPO-2275 Add convoluted test to trigger a very specific test condition
In the HippoAccessManager a fix was checked in on the
inprocessNodeReadAccess to clear a current node when the parent is
tested. However, with all the tests in the DocumentVariantShufflingTest
so far, still the problem would not be hit: This is because for
faceted navigation when the search is done against lucene, also the
authorization query is taken into account making sure only the readable
variants are returned in the search which will be tested for read access
afterward. *However*, only with unreadable results from the faceted
navigation query result we can trigger the state in which the
inprocessNodeReadAccess gives a problem when not removing the current
node id when checking the parent.

Hence, the only way to guarantee that the fix in the
inprocessNodeReadAccess will never be undone accidentally, is by
forcing an incorrect authorization query for the live user.

I had to open up some methods on the HippoInternalSession but added
a java doc comment that it is only meant for testing purposes which I
think is acceptable

Now the added test in the DocumentVariantShufflingTest really fails if
the fix in the HippoAccessManager is undone