REPO-2277 Authorization query incorrectly assumed JCR System Session
A user being marked as system session doesn't make it a JCR System Session
but makes the user a system user: A user required by the system to
function properly.

A mistakes is easily made. Improved javadoc, explicitly mentioning the
difference between a system user and a system session and added unit
test validate the authorization query is correct for a user that is
marked system