REPO-2278 rename confusing method (Internal)HippoSession.isSystemUser() to .isSystemSession()
HippoSession.isSystemUser() was too easily confused with User.isSystemUser() which means something completely different!
The latter indicates a non-human type of *user* which is required for the running system.
The first indicates a JCR system *session* having all privileges everywhere.

Related (internal) methods and properties within the HippoAccessManager and SessionImplHelper have been renamed as well
to make it explicitly clear they refer to a system *session* instead of a system *user*.