CMS-12319 make appPath configurable for plugins
Currently the appPath is derived from the perspective class name.
This has the undesired effect that the appPath is not easilly changed.
In addition, when a class is bootstrapped twice (with different config)
then we would end up with a duplicate appPath in the nav items.

I solved the issue by introducing a mixin that can be added to a plugin
The mixin requires a property frontend:appPath. This way we can query
for all nodes with this mixin to get all navItems. These items are
subsequently filtered in application code on the presence of the correct
hipposys:userrole. Therefore, any plugin that requires a specific role
should also be bootstrapped with that role. Otherwise the plugin is not
considered a navItem.

To make it easy to re-use logic I register the NavigationItemService in
the hippo service registry. That way the Perspective class can use the
service to get the navigation item that belongs to this perspective.
Note that the navigation item might be null if the user does not have
the required user role.

Translations are provided in the front-end. We do need to keep the
localization for custom perspective because the navapp will not have
translations for those. For developer convenience I already bootstrap
an empty root node for the english navigation item translations.