CMS-12270 Fix layout of picker dialog by making it resizable
The picker dialog requires an initial render step which is handled by
calling "Wicket.Window.current.resizer.restoreDatatableHeight();". This
function is only available when the dialog itself is resizable.
The link-picker dialog used in by CKEditor is already resizable, so it
makes sense to make the generic link-picker dialog resizable as well.