NAVAPP-320 Rename menu item ids to conform with back-end changes
In the back-end we no longer derive the menu item id from the
perspective class name. So we are now free to choose a more sensible id.
I chose to prefix all xm items with xm and aligned the suffix with the
appPath. So in general item ids of xm now follow the patten
xm-${appPath}.