REPO-2272 Catch the correct exception
Added integration/unit tests covering the DomainManager. One of the
tests failed because the wrong exception was caught in the
DomainsManagerImpl : it is a path and not item not found exception
that most be caught

Also note that a couple of the unit tests fail: this is on purpose: I
think the returned objects should be immutable and thus
immutable maps are expected. We either need to fix this or remove the
test assertions if not needed