CMS-12277 Read navigation resources from properties
Currently navapp resources are read from the repository. So resource
info needs to be bootstrapped via yaml files. A developer might not know
in advance where the CMS will be deployed, so he might not be able to
create the bootstrap data in advance. By using properties files we can
resolve this issue. By means of the HstServices it is easy to get hold
of the hst-config.properties and read the resources from there.

Navapp properties are namespaced by the prefix 'brx.navapp'. There are
currently 3 kinds of navapp resources:
brx.navapp.resource.navigationitems
brx.navapp.resource.login
brx.navapp.resource.logout

Each of these properties expects a value formatted as: $type, $url
where $type is REST or IFRAME and $url is the fully qualified url of the
resource.