CMS-12231 Hide the mask on Wicket.Window.close
There are several ways by which a modal window can get closed. In all
cases the navapp mask must get removed, because it was shown at the
moment the modal is shown (see show method). Doing the Hippo.hideMask()
call from java is error prone, because, as said there are multiple
code paths that can lead to the modal being closed.

Like Wicket.Window.show close is a prototype function that we can
specialize to make sure Hippo.hideMask() is called right after closing
the modal. In addition, I think it is also better to show the mask as
early as possible. So that is why I moved it into initialize.

Furthermore it is safe to assume that Hippo.showMask and Hippo.hideMask
are present. So I also removed the safety checks for the existence of
these methods.