NAVAPP-322 Add flag to NavLocation to push new item in history
When this flag is set to true a new history item will be pushed into the
browsers history. If the flag is not set, by default, the current
history item will be replaced. This assumes that client apps would have
a modern router which itself interacts with the browser history api.