CMS-12277 Return resources as Sets and check for duplicates
Resources should be unique, so it makes sense to return a Set instead of
a list to make this clear. Resources are read from properties, so it
could happen that two keys for e.g. login could have the same url. That
is most likely a configuration mistake. So to make sure these mistakes
are detected early an exception is thrown in case of duplicates.