CMS-12339 [Backport 13.4] Cover hippostd:directory as well and handle corner case correctly
Validate.isTrue throws an IllegalArgumentException if not true, breaking
the CMS UI completely. This is not allowed. If the condition is not met,
just return new JcrNodeModel("/") like also is being done for a
repository exception. new JcrNodeModel("/") results in a picker that
starts with the root folder instead of a specific folder being selected

(cherry picked from commit da7eb20467267eb7cc69767b8e593408b9b7d8ac)