CMS-12344 Use GET to logout
In the CMS iframe a jQuery.ajax call was done with the url of the Wicket
ajax Logout behavior. As a consequence the whole interaction is
considered an XHR request by the browser, so the browser uses an OPTIONS
request to auth0. Unfortunately auth0 responds in such a way that the
browser refuses to continue with the next redirect.

Solution (for now) will be to just force the browser location (of the
CMS iframe) to change to the logoutCallbackUrl of the Wicket behavior.
That way the browser will do a GET request and no problems with CORS
headers will occur. However, as a response to the logout call from the
navapp the CMS iframe must return a (resolved) Promise. This should of
course happen before the document location of the iframe is changed,
otherwise the promise will never be sent.