CMS-12308 Load top window if login is displayed in iframe
If a user can not (or no longer) be identified server-side the XM
login screen should be shown in the top window of the user's browser.

The server-side identity of a user can get lost in several ways:
- If the JSESSIONID cookie is deleted from the browser
- If the concurrent login filter detects a concurrent login

When that happens other servers (auth0 and brSM apps) are still able to
identify the user, because cookies on their domains are still there.
Ideally these cookies should also get removed from the browser when the
login screen is displayed, but that is not so easy without XM knowing
who the user is. So when the user chooses to use SSO she will not have
to fill in any credentials because the Auth0 cookie is still valid.

I think, for now it's a limitation that we can live with.